feat: use `forge fmt`

+++ src/ReservoirRouter.sol
@@ -20,8 +20,7 @@ contract ReservoirRouter is
     Multicall,
     SelfPermit
 {
-    constructor (address aFactory, address aWETH) PeripheryImmutableState(aFactory, aWETH)
-    {} // solhint-disable-line no-empty-blocks
+    constructor(address aFactory, address aWETH) PeripheryImmutableState(aFactory, aWETH) { } // solhint-disable-line no-empty-blocks
 
     function _addLiquidity(
         address aTokenA,
@@ -37,7 +36,8 @@ contract ReservoirRouter is
             rPair = factory.createPair(aTokenA, aTokenB, aCurveId);
         }
 
-        (uint256 lReserveA, uint256 lReserveB) = ReservoirLibrary.getReserves(address(factory), aTokenA, aTokenB, aCurveId);
+        (uint256 lReserveA, uint256 lReserveB) =
+            ReservoirLibrary.getReserves(address(factory), aTokenA, aTokenB, aCurveId);
         if (lReserveA == 0 && lReserveB == 0) {
             (rAmountA, rAmountB) = (aAmountADesired, aAmountBDesired);
             return (rAmountA, rAmountB, rPair);
@@ -65,7 +65,8 @@ contract ReservoirRouter is
         address aTo
     ) external payable returns (uint256 rAmountA, uint256 rAmountB, uint256 rLiq) {
         address lPair;
-        (rAmountA, rAmountB, lPair) = _addLiquidity(aTokenA, aTokenB, aCurveId, aAmountADesired, aAmountBDesired, aAmountAMin, aAmountBMin);
+        (rAmountA, rAmountB, lPair) =
+            _addLiquidity(aTokenA, aTokenB, aCurveId, aAmountADesired, aAmountBDesired, aAmountAMin, aAmountBMin);
 
         _pay(aTokenA, msg.sender, lPair, rAmountA);
         _pay(aTokenB, msg.sender, lPair, rAmountB);
@@ -95,15 +96,13 @@ contract ReservoirRouter is
     }
 
     /// @dev requires the initial amount to have already been sent to the first pair
-    function _swapExactForVariable(
-        uint256 aAmountIn,
-        address[] memory aPath,
-        uint256[] memory aCurveIds,
-        address aTo
-    ) private returns (uint256 rFinalAmount) {
+    function _swapExactForVariable(uint256 aAmountIn, address[] memory aPath, uint256[] memory aCurveIds, address aTo)
+        private
+        returns (uint256 rFinalAmount)
+    {
         require(aAmountIn <= type(uint112).max, "RR: AMOUNT_IN_TOO_LARGE");
         int256 lAmount = int256(aAmountIn);
-        for (uint i = 0; i < aPath.length - 1; ) {
+        for (uint256 i = 0; i < aPath.length - 1;) {
             (address lInput, address lOutput) = (aPath[i], aPath[i + 1]);
             (address lToken0,) = ReservoirLibrary.sortTokens(lInput, lOutput);
             address lTo = i < aPath.length - 2
@@ -111,10 +110,14 @@ contract ReservoirRouter is
                 : aTo;
             lAmount = lInput == lToken0 ? int256(lAmount) : -int256(lAmount);
 
-            lAmount = int256(IReservoirPair(ReservoirLibrary.pairFor(address(factory), lInput, lOutput, aCurveIds[i])).swap(
-                lAmount, true, lTo, new bytes(0)
-            ));
-            unchecked { i += 1; }
+            lAmount = int256(
+                IReservoirPair(ReservoirLibrary.pairFor(address(factory), lInput, lOutput, aCurveIds[i])).swap(
+                    lAmount, true, lTo, new bytes(0)
+                )
+            );
+            unchecked {
+                i += 1;
+            }
         }
         // lAmount is guaranteed to be positive at this point
         rFinalAmount = uint256(lAmount);
@@ -127,7 +130,12 @@ contract ReservoirRouter is
         uint256[] calldata aCurveIds,
         address aTo
     ) external payable returns (uint256 rAmountOut) {
-        _pay(aPath[0], msg.sender, ReservoirLibrary.pairFor(address(factory), aPath[0], aPath[1], aCurveIds[0]), aAmountIn);
+        _pay(
+            aPath[0],
+            msg.sender,
+            ReservoirLibrary.pairFor(address(factory), aPath[0], aPath[1], aCurveIds[0]),
+            aAmountIn
+        );
         rAmountOut = _swapExactForVariable(aAmountIn, aPath, aCurveIds, aTo);
         require(rAmountOut >= aAmountOutMin, "RR: INSUFFICIENT_OUTPUT_AMOUNT");
     }
@@ -139,7 +147,7 @@ contract ReservoirRouter is
         uint256[] memory aCurveIds,
         address aTo
     ) private {
-        for (uint i = 0; i < aPath.length - 1; ) {
+        for (uint256 i = 0; i < aPath.length - 1;) {
             (address lInput, address lOutput) = (aPath[i], aPath[i + 1]);
             (address lToken0,) = ReservoirLibrary.sortTokens(lInput, lOutput);
             // PERF: Can avoid branching on every iteration by moving the last step outside of the for loop
@@ -153,7 +161,9 @@ contract ReservoirRouter is
                 lAmount, false, lTo, new bytes(0)
             );
 
-            unchecked { i += 1; }
+            unchecked {
+                i += 1;
+            }
         }
     }
 
@@ -167,7 +177,12 @@ contract ReservoirRouter is
         rAmounts = ReservoirLibrary.getAmountsIn(address(factory), aAmountOut, aPath, aCurveIds);
         require(rAmounts[0] <= aAmountInMax, "RR: EXCESSIVE_INPUT_AMOUNT");
 
-        _pay(aPath[0], msg.sender, ReservoirLibrary.pairFor(address(factory), aPath[0], aPath[1], aCurveIds[0]), rAmounts[0]);
+        _pay(
+            aPath[0],
+            msg.sender,
+            ReservoirLibrary.pairFor(address(factory), aPath[0], aPath[1], aCurveIds[0]),
+            rAmounts[0]
+        );
         _swapVariableForExact(rAmounts, aPath, aCurveIds, aTo);
     }
 }
